Fix hoverProperty of first tile to not blink
Problem: When adding a large number of images to the
ImageTileScrollPane, the hoverProperty of the first and second tiles
flickers between true and false while mouse cursor on the tiles.

Cause: When an image is loaded into a tile, the tile is displayed at
(0,0) and covers the first and second tiles.
Since the tile move to the correct layout in a moment, it is invisible
to humans but intercepts the hover events of the mouse.

Solution: Do not display tiles that have not yet been moved to the
correct position.